+++ packages/orq-rc/src/models/operations/modelcreateawsbedrock.ts
@@ -15,6 +15,8 @@ export type ModelCreateAwsBedrockRequestBody = {
   authMode: string;
   autorouterId?: string | undefined;
   autorouterVersion?: string | undefined;
+  cacheReadCost?: number | undefined;
+  cacheWriteCost?: number | undefined;
   description?: string | undefined;
   displayName: string;
   hasReasoning?: boolean | undefined;
@@ -79,6 +81,8 @@ export type ModelCreateAwsBedrockRequestBody$Outbound = {
   auth_mode: string;
   autorouter_id?: string | undefined;
   autorouter_version?: string | undefined;
+  cache_read_cost?: number | undefined;
+  cache_write_cost?: number | undefined;
   description?: string | undefined;
   display_name: string;
   has_reasoning?: boolean | undefined;
@@ -112,6 +116,8 @@ export const ModelCreateAwsBedrockRequestBody$outboundSchema: z.ZodType<
   authMode: z.string(),
   autorouterId: z.string().optional(),
   autorouterVersion: z.string().optional(),
+  cacheReadCost: z.number().optional(),
+  cacheWriteCost: z.number().optional(),
   description: z.string().optional(),
   displayName: z.string(),
   hasReasoning: z.boolean().optional(),
@@ -139,6 +145,8 @@ export const ModelCreateAwsBedrockRequestBody$outboundSchema: z.ZodType<
     authMode: "auth_mode",
     autorouterId: "autorouter_id",
     autorouterVersion: "autorouter_version",
+    cacheReadCost: "cache_read_cost",
+    cacheWriteCost: "cache_write_cost",
     displayName: "display_name",
     hasReasoning: "has_reasoning",
     inputCost: "input_cost",